Add option for strict placeholder parsing
Strict parsing is useful during development
and to verify if all placeholders are present
on a deploy target.

It is counter productive, when building the project
on a build server, where some placeholder accessor targets
might not exist.

Therefore we provide both ways to dump settings and let
the user decide which one to use, defaulting to not
throwing exceptions.