[FEATURE] Add option to dump config as dynamic PHP code
To satisfy the need to directly react e.g. on environment
variable changes, it is required that the exported configuration
array directly pulls environment variables out of the environment.

By doing so, the exported configuration can stay the same
(e.g. in a container), but the environment can still change
and the Configuration will adapt.

While doing so, we create PHP code for the other placeholders
as well (global, const, conf) to be consistent.

Since this approach has slightly decreased runtime performance,
this is done optionally, thus only when users decide to use the
introduced --code option for dumping configuration.