[FEATURE] Add a way to import a resource from env var
In some cases, looking at you Docker, it is tedious to provide
a crafted configuration file in an environment, but simple to
provide environment variables that hold an environment name.

For this a special import type is now added, which allows specifying
an environment variable and derive a configuration file from the
environment variable's value.

This allows complex setups. In general it is recommended though to
use the most simple approach possible that matches the
environment requirements at hand.