[TASK] Only read override settings file if it has config
In out setup we don't use the `defaultConfiguration` step but use the use the `SetupConfigurationAction` to set `customOverrideSettings`.

Because of the missing `defaultConfiguration` step the `override.settings.yaml` was not created before the `SetupConfigurationAction` got executed. In this case `readConfig()` throws an error that the file does not exist.

Therefore we have to check if the file has config before reading it.